chat/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
backports.asyncio.runner==1.2.0
certifi==2025.7.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
defusedxml==0.7.1
Django==5.1.7
django-extensions==4.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.3
dotenv==0.9.9
exceptiongroup==1.3.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
msgpack==1.1.0
oauthlib==3.3.1
packaging==25.0
pluggy==1.6.0
psycopg==3.2.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.2
PyJWT==2.9.0
pyOpenSSL==25.0.0
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-django==4.11.1
python-dotenv==1.0.1
python3-openid==3.2.0
redis==5.2.1
requests==2.32.4
requests-oauthlib==2.0.0
service-identity==24.2.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.5.0
whitenoise==6.9.0
zope.interface==7.2
